updated Rectangle, Square, Quadrilateral

+++ src/shapes/Quadrilateral.java
@@ -3,7 +3,6 @@ package shapes;
 public abstract class Quadrilateral extends Shape implements Measurable{
 
     protected double length;
-
     protected double width;
 
     public Quadrilateral(double length, double width) {
@@ -11,12 +10,15 @@ public abstract class Quadrilateral extends Shape implements Measurable{
         this.width = width;
     }
 
+    // Getters
     public double getLength() {
         return this.length;
     }
     public double getWidth() {
         return this.width;
     }
-    public abstract void setLength(int length);
-    public abstract void setWidth(int width);
+
+    // Setters
+    public abstract void setLength(double length);
+    public abstract void setWidth(double width);
 }
