+++ src/shapes/Square.java
@@ -22,12 +22,12 @@ public class Square extends Quadrilateral implements Measurable {
     }
 
     @Override
-    public void setLength(int length) {
+    public void setLength(double length) {
         this.length = length;
     }
 
     @Override
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
